pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
    "python-dateutil>=2.9.0.post0",
]
